requirements.txt
tweepy==3.10.0
colorama==0.4.4
requests_html
Pillow